tsdb/x/tsdb/types/messages_timeseries_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"github.com/supsi-dacd-isaac/cosmos-apps/tsdb/testutil/sample"
)

func TestMsgCreateTimeseries_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateTimeseries
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateTimeseries{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateTimeseries{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdateTimeseries_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateTimeseries
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateTimeseries{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateTimeseries{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeleteTimeseries_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeleteTimeseries
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeleteTimeseries{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeleteTimeseries{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
